try for first test

+++ vertexShader.glsl
@@ -1,8 +1,8 @@
                               # vertexShader.glsl
 #version 330 core
 
-layout (location = 0) in int x;
-layout (location = 1) in int y;
+layout (location = 0) in float x;
+layout (location = 1) in float y;
 
 void main() {
 	gl_Position = vec4( x, y, 0.0, 1.0 );
